+++ R/funKerasCensus.R
@@ -13,7 +13,7 @@
 #' @param nfactors (character) number of factor variables
 #' @param nnumericals (character) number of numerical variables
 #' @param cardinality (character) cardinality
-#' @param cachedir cachedie
+#' @param cachedir cachedir
 #' @param k number of repeats
 #' @param prop split proportion. Default: \code{c(3/5,1)}.
 #' @param verbosity verbsity. Default: 0
@@ -127,7 +127,7 @@ predMlCensus <- function(x = NULL,
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #' cfg <- getModelConf(list(model="dl"))
 #' x <- matrix(cfg$defaults, nrow=1)
@@ -223,7 +223,7 @@ predDlCensus <- function(x = NULL,
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #' target <- "age"
 #' task.type <- "classif"
@@ -321,7 +321,7 @@ getMlConfig <- function(target,
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #' ## The following code was used to evaluate the results in the book
 #' ## "Hyperparameter Tuning for Machine and Deep Learning with R - A Practical Guide"
@@ -462,7 +462,7 @@ evalParamCensus <- function(
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #' runNrMl <- list("15")
 #' runNrDl <- list("28")