+++ R/funKerasMnist.R
@@ -21,7 +21,7 @@
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #'
 #' library("SPOTMisc")
@@ -130,7 +130,7 @@ getMnistData <- function(kerasConf) {
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #'
 #' library("SPOTMisc")
@@ -386,7 +386,7 @@ evalKerasMnist <-
 #' ### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
 #' ### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
 #' ### June 2021. http://arxiv.org/abs/2105.14625.
-#' PYTHON_RETICULATE = FALSE
+#' PYTHON_RETICULATE <- FALSE
 #' if(PYTHON_RETICULATE){
 #'
 #' library("SPOTMisc")